WIP Refactor graph_algorithms module code into submodules
