+++ src/algorithms/network_flow.py
@@ -1,0 +1,10 @@
+from src.notation import Network, Graph
+
+
+class NetworkFlowAlgorithms:
+    def __init__(self, N: Network):
+        self.network = N
+        self.residual_network = N.copy()
+
+    def run(self) -> Graph:
+        pass
